Break import/export into independent components
