+++ wallet-server-ui/src/app/component/import/import.component.scss
@@ -1,0 +1,37 @@
+.import {
+  .row {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+
+    margin-bottom: 0.5rem;
+
+    label {
+      margin-right: 0.5rem;
+    }
+    button {
+      flex: 1;
+      margin-bottom: 0;
+    }
+  }
+  .radio-row {
+    .import-type {
+      margin-bottom: 0.5rem;
+      mat-radio-button {
+        margin-right: 1rem;
+      }
+    }
+  }
+  .import-textarea {
+    margin-bottom: -0.75rem;
+  }
+  .expansion-panel {
+    margin-bottom: 0.75rem;
+  }
+  .passphrase-input {
+    
+  }
+  .wallet-name-input {
+    
+  }
+}
